The list of themes is created in a separate file
